make saga snapshot storage factory disposable so it can clean things up
